requirements.txt
openai
pyyaml
mdpdf
streamlit
watchdog